backend/requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.2
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.0
extensions==0.4
git-filter-repo==2.47.0
iniconfig==2.1.0
packaging==25.0
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.5
pytest-django==4.11.1
python-dotenv==1.1.0
python-environ==0.4.54
sqlparse==0.5.3
tzdata==2025.2
